requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
langchain-core
streamlit
langchain-google-genai
langchain-ollama

